- Moved MenuLogic to Reservation
