+++ ReservationSystem/PresentationLayer/Menu/MenuPresentation.cs
@@ -1,0 +1,6 @@
+using ReservationSystem;
+
+public class MenuPresentation : View
+{
+    
+}
